Extract attestation integration tests into seperate crate
Fixes http://b/326052966
