+++ oak_attestation_integration_tests/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "oak_attestation_integration_tests"
+version = "0.1.0"
+authors = ["Juliette Pluto <julsh@google.com>"]
+edition = "2021"
+license = "Apache-2.0"
+
+[dev-dependencies]
+oak_restricted_kernel_sdk = { workspace = true, features = [
+  "mock_attestation"
+] }
+oak_attestation = { workspace = true }
+oak_attestation_verification = { workspace = true }
+oak_proto_rust = { workspace = true }
