+++ oak_attestation_integration_tests/README.md
@@ -1,0 +1,4 @@
+# Oak Attestation Integration Tests
+
+Tests that cover the full workflow of generating an attestation to then
+verifying it.
